Allow `format_pyproject` to be called with text
Before this change, `format_pyproject` needed to be called with a file
path, instead of the file contents.

This change tries to make it easier to use `format_pyproject` from a
programmatic point of view. Authors of a Python script or library not
necessarily have an existing file with the TOML contents that needed to
be formatter, but they usually have access to the string contents
directly.